Add command.Handler as single source of truth for device commands
- Introduce internal/command package with ExecuteRPC + ApplySharedAttributes
- Map RPC methods (reboot, shutdown, locate, setLocationInterval, setStatusInterval)
  and shared attributes (locationIntervalSeconds, statusIntervalMinutes)
- Server implements SessionLookup.SendToDevice
- Debug REST becomes a thin wrapper over the same handler

This prepares the path for ThingsBoard Gateway MQTT RPC while keeping
the existing debug API working.

+++ internal/protocol/commands.go
@@ -67,11 +67,11 @@ func BuildUploadInterval(seconds uint16) []byte {
 }
 
 // BuildStatusInterval sets status packet interval via 0x13 (minutes).
-func BuildStatusInterval(minutes byte) []byte {
+func BuildStatusInterval(minutes int) []byte {
 	if minutes == 0 {
 		minutes = 1
 	}
-	return []byte{0x78, 0x78, 0x02, MsgStatus, minutes, 0x0D, 0x0A}
+	return []byte{0x78, 0x78, 0x02, MsgStatus, byte(minutes), 0x0D, 0x0A}
 }
 
 // BuildLocate requests an immediate position upload (0x80).
